requirements.txt
streamlit==1.31.0
pandas==2.2.0
plotly==5.18.0 
requests==2.31.0 
python-dotenv==1.0.0 
matplotlib==3.8.2 
numpy==1.26.3 
fpdf2==2.7.6
